Move view_course_access functionality to sole caller lms/mobile_api.
